+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/annotators/DefaultAnnotator.kt
@@ -21,6 +21,24 @@ class DefaultAnnotator : BaseAnnotator() {
 
     override fun getKeywordType(element: PsiElement): TextAttributesKey? {
         var type: TextAttributesKey? = null
+        val supportLanguages = listOf(
+            "kotlin",
+            "ObjectiveC",
+            "Dart",
+            "go",
+            "JAVA",
+            "JavaScript",
+            "php",
+            "python",
+            "rust",
+            "typescript"
+        )
+
+        try {
+          if (supportLanguages.map { it.lowercase() }.contains(element.language.id.lowercase())) return null
+        } catch(_: Exception) {
+        }
+
         when (element.text) {
             "new", "return" -> type = SECONDARY_KEYWORD
             "if", "else", "switch", "case", "default", "break", "continue", "assert" -> type = SECONDARY_KEYWORD