+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # vscode-theme Changelog
 
 ## [Unreleased]
+### Changed
+- Default Annotation - Decommissioning Default Annotator due to Issue #38
+- Typescript await keyword inconsistency.
 
 ## [1.7.3]
 ### Fixed